👌 IMPROVE: Improve course finder and clear mock data
