--- src/components/CourseFinder/item/index.jsx
@@ -1,2 +1,0 @@
-export { default as CourseItemMain } from './CourseItemMain'
-export { default as CourseItemSub } from './CourseItemSub'
