+++ src/components/Favourites/index.jsx
@@ -1,2 +1,3 @@
 // eslint-disable-next-line import/prefer-default-export
 export { default as FavouritesContainer } from './FavouritesContainer'
+export { default as FavoriteToggle } from './FavoriteToggle'
