+++ src/components/Timetable/index.jsx
@@ -1,3 +1,4 @@
 export { default as TimetableContainer } from './TimetableContainer'
 export { default as TimetableCourseItem } from './TimetableCourseItem'
 export { default as TimetableLayout } from './TimetableLayout'
+export { default as TimetableSelector } from './TimetableSelector'
